Fixed manager parameters and removed "Half modal" code
Removed tableview in parameters for method, it was used at one time and
not needed anymore
Changed name of setDataFilterNames to setFiltersWithFilterNames
Removed all the “half modal” code which means observer gone, etc.
I tried a really long question and saw I wasn’t handling positioning
for the text field in that case.  I added a new parameter on the text
field and made it so the label sized to the text.