Example of subclass using methods
Here’s the example now using the two methods. The hierarchy/complicated
one makes an extra call to setCurrentManagerSettings. The
filterviewcontroller keeps track of the current index and sets the
settings for that one. I’m not sure if that makes it more complicated
but then the filterviewcontroller keeps track of the entire manager
array and the subclass is oblivious to it.