Made look the same for both modals
The tableviews for the modal pieces were a style of grouped. This made
the text field and tableview very different in style. The tableview
subtitle text as also in caps, so the willDisplayHeader code is to
override that with the real question rather than caps. I also made the
background of the text field modal the same as the tableview with a
method for a sub class to change the background for the modals.