Fixed wording for text field and added device orientation code
Alex wanted the wording changed sp I changed it in the .strings file to
show her what it would look like.

This also includes the fix to properly show the text field/ui label on
the text modal whenever the device is rotated while the modal is up.
The view controller would rotate it but the text field would be too low
and underneath the keyboard. So I made set methods instead of creation
methods to reuse the code, added observing for the device orientation
and made cancel/save remove observing.