Changes to string files
Since this needs to be a cocoa pod, the classes should allow for
overriding so that a strings file in the implementation can be used
instead of having to update multiple files. The same terminology should
be used but now the files can be overwritten if needed. I moved the
survey, interactions and label strings into an example file to be put
in the example project for the pod.