Changed number ids to string ids added logging on example
Strings are needed for mission hub anyway from what I can tell. The
model has a nsnumber for the id but the filter string will be built
using strings, so it’s better to have the unique id as a string. The
NSUUID class can convert the id to a string.

I also added a method printing in the example so the developer could
see the filter returned from the manager.